+++ Idp/__init__.py
@@ -5,3 +5,4 @@ from .Substitute import Done
 from .Implicant  import Done
 from .Idp_to_Z3  import Done
 from .Run        import Done
+from .Problem    import Done